Rework app launch to ensure DB isn't accessed before ready.
